pyproject.toml
[project]
name = "intervals-icu-mcp"
version = "1.0.0"
description = "Model Context Protocol server for Intervals.icu API integration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.4",
    "httpx>=0.28.1",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
intervals-icu-mcp = "intervals_icu_mcp.server:main"
intervals-icu-mcp-auth = "intervals_icu_mcp.scripts.setup_auth:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "respx>=0.22.0",
    "ruff>=0.14.0",
    "pyright>=1.1.406",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["E501"]  # Line length enforced by ruff format instead

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files
"src/intervals_icu_mcp/server.py" = ["E402"]  # Imports must come after FastMCP initialization

[tool.pyright]
strict = ["src"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
